[ATfL] Stop using timezone database in our build of libc++ (#77)
It's a cherry-pick from the arm-software branch.

The format of the /etc/localtime file can be different across the
different Linux installations, and tzdb routines in libc++ currently
cannot cope with that. In order to improve portability and to avoid any
struggle with failing tzdb test cases, we should disable it.