Add newlib-nano as multilib overlay package (#60)
See
https://github.com/ARM-software/LLVM-embedded-toolchain-for-Arm/issues/628
for context.

This enables newlib-nano as a multilib. I didn't find any release
scripts, so I don't know how to add it to the CI such that it also
generates an overlay package for newlib-nano on release.

Note: this also fixes an issue with the current newlib builds where they
don't get built correctly for anything above ARMv6 as the `-march`,
`-mfloat-abi` and `-mfpu` arguments weren't being passed on to the
newlib compilation configuration. A little unsure why this hasn't popped
up before...