[ATfL] Do not use libzstd.a when building shared runtime libs (#130)
At the shared_lib_build stage we do not need ZSTD as the main goal of
this stage is to provide shared runtime libraries, while ZSTD is being
used at compile time. Sadly, there is no way (yet) to eliminate
remaining attempts to build other things at this stage, and some of this
stuff is linked against libzstd. Unfortunately, on some Linux
distributions, those static variants of libraries that we are using to
build ATfL were not built with -fPIC, resulting in linker failures. The
only way to prevent this is to override one of the COMMON_CMAKE_FLAGS
and disable the use of ZSTD at this stage.