[Downstream change] Implement Arm Toolchain ID (#126)
Since there are potentially different versions of the toolchain that can
be built (e.g. Arm Toolchain for Embedded and Arm Toolchain for Linux),
we need a way to uniquely identify the toolchain from the version
output, and differentiate from the upstream version. This can be done by
appending an additional message:
`Arm Toolchain ID: tag.`

The contents of the tag will need to be generated by the product's build
script. This patch add the field to version output, which can be set by
the CMake option ARM_TOOLCHAIN_ID when building the toolchain.

Clang and clang-format have their own version functions, whereas the
other LLVM tools use a shared utility function. This patch appends the
new message into both.

---

Downstream issue: #124