[ATfE] Set Arm Toolchain ID (#127)
In order to uniquely identify ATfE builds, this patch populates the new
Arm Toolchain ID with an auto-generated tag.

This tag contains both the product code (E for Embedded), as well as a
unique identifier for the build. In a Jenkins environment, this will be
based on the build URL and build number.