[ATfE] Restrict variable replacement in check-picolibc script (#175)
The `LIT_OPTS` environment variable should be evaluated at runtime, not
during configuration. The `@ONLY` option can be used to limit
replacement to only the variables that need replacing during
configuration.

Additionally, compiling the tests has been added as a dependancy of
building the library so that it can be done in parallel.