[ATfE] Don't delete temporary semihosting-features file (#214)
The NamedTemporaryFile used to create the `:semihosting-features` file
for FVP testing will automatically attempt to delete itself when closed.
Since this file needs to persist in order to replace any existing copy
of the file, the auto deletion must be disabled.

The `os.rename` operation has also been changed to `os.replace` since
renaming to an existing file will throw an error on some operating
systems such as Windows.