[ATfE] v8-M library naming and selection issues for strict alignment builds. (#250)
We don’t have a -mno-unaligned-access specific variant for the v8M libraries. For v8-M with strict aligned access, it’s more appropriate to use the v7-M variants, as they include a version with strict alignment support. It was a deliberate decision to provide strict-alignment variants only for a subset of architectures.

To support this, both -mno-unaligned-access and -munaligned-access options must be passed into the multilib selection logic. This aligns with the approach in https://github.com/llvm/llvm-project/pull/134099/, which also considers the target architecture when determining the alignment access policy. Since Armv6-M does not support unaligned accesses, test changes are needed to reflect the updated behaviour when the target architecture is taken into account.

This patch fixes below issues.

1. The naming of the v8-M libraries should explicitly include
‘unaligned’ to reflect their behaviour.

2. When using -mno-unaligned-access, the multilib system should instead
select the v7-M variant.

3. Armv6m test changes are needed to reflect the updated behaviour when
the target architecture is taken into account.