[ATfL] Add package documentation (#248)
This patch adds the ATfL package documentation which should land into:

1. Our on-line page(s)

2. Our release .deb packages

3. Our release .rpm packages

The markdown documents introduced by this patch have landed into the
arm-software/linux/docs directory. This follows the same guidance
that has been already agreed for the ATfE product (see [1]).

Additionally, an index.md file has been created which follows the
convention already observed in the LLVM's documentation directories
(e.g., [2]).

The files from the arm-software/linux/docs directory are being copied
into the arm/docs directory of the tarball and this is happening in
the build.sh script also modified by this patch. Whatever lands into
that directory, should act as a basis for further rendering of the
on-line page(s) and the PDF document(s) in our CI pipelines.

[1]
https://github.com/arm/arm-toolchain/blob/arm-software/arm-software/docs/README.md

[2]
https://github.com/arm/arm-toolchain/blob/arm-software/flang/docs/index.md