[ATfE] Only check machine name when run_fvp.py runs (#320)
Since the run_fvp.py script is imported into other scripts, the machine
name check can be carried out on machines where FVP testing is not
supported and disabled.

This patch moves the check inside the run function, so that it the
machine name only needs to be matched when FVP tests are actually being
executed.