[ATfE] Disable package tests on macOS (#327)
The package tests rely on CMake to extract the package, however the
command used only supports tar and zip archives. Until an alternative
solution is found to test the dmg archive produced on macOS, the package
tests can instead be disabled to prevent a failure.

The `test.sh` script will still attempt to build the
`check-package-llvm-toolchain` target, so a dummy target is created to
avoid an unrecognized target error.