[ATfE] Re-disable aarch64 nofp variants for newlib (#329)
https://github.com/arm/arm-toolchain/pull/314 should have only enabled
aarch64 nofp variants for LLVM-libc builds. However it also enabled the
variants for newlib builds, which were disabled since newlib is
similarly unable to build those variants (see
https://github.com/arm/arm-toolchain/pull/60#issuecomment-2648640369).
While there is now a fix for the LLVM-libc build, there is yet no fix
for newlib so they need to remain disabled for that selection.

This patch sets the variants to only build for picolibc and llvmlibc,
but not newlib.