[ATfE] Fix and re-enable package test on MacOS (#339)
The package test implementation curretly relies on using the CMake `tar`
command to unpack the toolchain for testing. As this CMake command is
not able to handle .dmg image files, the package test was disabled
temporarily on MacOS platforms.
This change implements an alternative way to unpack the toolchain .dmg
file for testing on MacOS. It introduces a new `extract_dmg.sh` script
that makes use of the `hdiutil` program to extract the contents
correctly. With the fix in place, this change also re-enables the
package tests on MacOS platforms.