[ATfL] build.sh modification to generate manpages including one for armclang (#353)
Unfortunately, currently the flang manpages cannot be generated despite
ongoing effort to make it possible.

The list of dependencies had to be extended, as the documentation
processing requires the dot command which is provided by the graphviz
package.

The modulefiles did not have to be modified, they already handled the
MANPATH variable.