[ATfE] pacret_bti variants shouldn't be built with pacbti
In multilib.json the flags for these library variants don't have
pacbti, so if we compile with e.g. -mcpu=cortex-m55
-mbranch-protection=standard we select the pacret_bti variant and
end up executing an undefined bxaut instruction. Fix this by not
building with pacbti, which means -mbranch-protection=standard causes
instructions to be used that are forward-compatible with pacbti but
are nop without it.