Configure code owners to auto-assign PRs to relevant teams (#403)
This change updates the contents of `.github/CODEOWNERS`, replacing the
orignal upstream content with a mapping to the teams working on
arm/arm-toolchain.

By default, both the ATfE and ATfL teams are assigned as reviewers. This
is particularly revelant for downstream changes to LLVM code, where the
downstream contribution policy must be enforced. Changes to each of the
toolchain's specific directories are assigned to the relevant team only.