Check review approvals in downstream changes (#408)
Currently, pull requests only require a single approved review in order
to merge. For downstream changes, we would prefer both product teams to
review the changes first, but GitHub does not have the functionality to
set different merge requirements depending on the paths of the changed
files.

However, since we have a script that already runs on downstream changes
to check it meets certain requirements, this could also check the review
state and which teams have approved. This patch adds this additional
check, getting the review information from the existing query and using
the API to get the members of each team. If a review from a team is
missing, the check will fail which will prevent merging. Once both have
approved, the check will pass allowing merging.

The script automatically runs on modifications to the pull request, so
it is additionally set to trigger when a pull request is reviewed. The
`pull_request_review` trigger does not have a `branches` filter, so the
job is set to run conditionally instead.