[ATfE] Move errno from semihost to libc and add time utils (#436)
Previously, we used an external errno in LLVM-libc, which had to be
defined in the semihosting library. However, LLVM-libc does provide an
option to use a single-threaded errno implementation on request, so for
consistency with other C libraries, we move errno from `libsemihost.a`
to `libc.a`.

Also add new functions to allow clock().