[ATfE] Changes in library variants build for code size (#429)
- All the M-profile variants have their build type changed to
`minsizerelease`. Exceptions apply to Armv8.1-M Main. See below.
- For Armv8.1-M Main, the following changes apply:
- All the PACBTI variants have their build type changed to
`minsizerelease`.
  - Ditto for all the soft nofp nomve variants.
- For the `unaligned` variants that don't fall under the previous
categories, a new version is created so that two versions now coexist:
one with `release` as the build type, and another with `minsizerelease`.
- All variants whose build type is `minsizerelease` are renamed to
include the "size" segment in their names. This includes not only
M-profile but also Armv4t and Armv5te.
- Update tests and add ones that were missing, for example for Armv4t
and Armv5te.