Disable review approval portion of downstream change check (#449)
The default access token used in GitHub Actions workflows does not have
the required permissions to see the list of team members, which exists
outside of the arm-toolchain repo at the arm org level. As a result,
attempting to check which teams have reviewed a pull request will
currently cause an error.

For now, remove the check until it can be reworked. Since there will no
longer be a hard gated check on whether a pull request has been
sufficiently reviewed, the automated comment has been updated to remind
to wait for both teams before merging.